Added Example Code configured for our motor config
Summary:

This is an example code representing how to perform basic autonomous movement with P controller on driving and turning. This has been taken from the first external samples, and then configured to use our hardware mapping, with adjusted formatting and comments.

Next Steps:

Adjust to the proper pathing, walk Britney how to do this, note, this may be unnecessary depending on Deviation's current code progress on 12/9/2023, and what is/is not working.